// frontend/packages/zobi-ui-chart-controls/src/types.ts
type AnyDict = Record<string, any>;
// ...
export type ColumnMeta = Omit<Column, 'id'> & {
  id?: number;
} & AnyDict;
// ...
export enum SortSeriesType {
  Name = 'name',
  Max = 'max',
  Min = 'min',
  Sum = 'sum',
  Avg = 'avg',
}

export type LegendPaddingType = {
  top?: number;
  bottom?: number;
  left?: number;
  right?: number;
};

export type SortSeriesData = {
  sort_series_type: SortSeriesType;
  sort_series_ascending: boolean;
};